feat: add caching to the platform helper and rename
